+++ app/graphsensemodel.py
@@ -1,4 +1,3 @@
-from enum import Enum
 
 
 def byte_to_hex(bytebuffer):
@@ -47,8 +46,8 @@ class AddressSummary(object):
 # CASSSANDRA TABLES
 class ExchangeRate(object):
     def __init__(self, d):
-        self.usd = d['eur']
-        self.eur = d['usd']
+        self.eur = d['eur']
+        self.usd = d['usd']
 
 
 class Statistics(object):
@@ -164,6 +163,12 @@ def compute_balance(total_received_satoshi, total_spent_satoshi,
     return balance
 
 
+def compute_exchanged_value(value, exchange_rate):
+    return Value(value,
+                 round(value*exchange_rate.eur*1e-8, 2),
+                 round(value*exchange_rate.usd*1e-8, 2))
+
+
 class AddressTransactions(object):
     def __init__(self, row, rates):
         self.address = row.address
@@ -206,15 +211,19 @@ class Cluster(object):
 
 
 class AddressIncomingRelations(object):
-    def __init__(self, row):
+    def __init__(self, row, exchange_rate):
         self.dstAddressPrefix = row.dst_address_prefix
         self.dstAddress = row.dst_address
-        self.srcCategory = Category(row.src_category)
         self.estimatedValue = Value(row.estimated_value.satoshi,
                                     round(row.estimated_value.eur, 2),
                                     round(row.estimated_value.usd, 2)).__dict__
         self.srcAddress = row.src_address
         self.noTransactions = row.no_transactions
+        self.srcBalance = compute_balance(row.src_properties.total_received,
+                                          row.src_properties.total_spent,
+                                          exchange_rate)
+        self.srcTotalReceived = compute_exchanged_value(row.src_properties.total_received,
+                                                        exchange_rate)
         self.srcProperties = AddressSummary(row.src_properties.total_received,
                                             row.src_properties.total_spent)
 
@@ -226,8 +235,8 @@ class AddressIncomingRelations(object):
                 "nodeType": "address",
                 "received": self.srcProperties.totalReceived,
                 "balance": (self.srcProperties.totalReceived -
-                            self.srcProperties.totalSpent),  # satoshi
-                "category": self.srcCategory.name}
+                            self.srcProperties.totalSpent)  # satoshi
+                }
         return node
 
     def toJsonEdge(self):
@@ -237,17 +246,31 @@ class AddressIncomingRelations(object):
                 "estimatedValue": self.estimatedValue}
         return edge
 
+    def toJson(self):
+        return {
+            "id": self.id(),
+            "nodeType": "address",
+            "received": self.srcTotalReceived.__dict__,
+            "balance": self.srcBalance.__dict__,
+            "noTransactions": self.noTransactions,
+            "estimatedValue": self.estimatedValue
+        }
+
 
 class AddressOutgoingRelations(object):
-    def __init__(self, row):
+    def __init__(self, row, exchange_rate):
         self.srcAddressPrefix = row.src_address_prefix
         self.srcAddress = row.src_address
-        self.dstCategory = Category(row.dst_category)
         self.estimatedValue = Value(row.estimated_value.satoshi,
                                     round(row.estimated_value.eur, 2),
                                     round(row.estimated_value.usd, 2)).__dict__
         self.dstAddress = row.dst_address
         self.noTransactions = row.no_transactions
+        self.dstBalance = compute_balance(row.dst_properties.total_received,
+                                          row.dst_properties.total_spent,
+                                          exchange_rate)
+        self.dstTotalReceived = compute_exchanged_value(row.dst_properties.total_received,
+                                                        exchange_rate)
         self.dstProperties = AddressSummary(row.dst_properties.total_received,
                                             row.dst_properties.total_spent)
 
@@ -259,8 +282,8 @@ class AddressOutgoingRelations(object):
                 "nodeType": "address",
                 "received": self.dstProperties.totalReceived,
                 "balance": (self.dstProperties.totalReceived -
-                            self.dstProperties.totalSpent),  # satoshi
-                "category": self.dstCategory.name}
+                            self.dstProperties.totalSpent)  # satoshi
+                }
         return node
 
     def toJsonEdge(self):
@@ -269,6 +292,16 @@ class AddressOutgoingRelations(object):
                 "transactions": self.noTransactions,
                 "estimatedValue": self.estimatedValue}
         return edge
+
+    def toJson(self):
+        return {
+            "id": self.id(),
+            "nodeType": 'address',
+            "received": self.dstTotalReceived.__dict__,
+            "balance": self.dstBalance.__dict__,
+            "noTransactions": self.noTransactions,
+            "estimatedValue": self.estimatedValue
+        }
 
 
 class ClusterSummary(object):
@@ -279,10 +312,9 @@ class ClusterSummary(object):
 
 
 class ClusterIncomingRelations(object):
-    def __init__(self, row):
+    def __init__(self, row, exchange_rate):
         self.dstCluster = str(row.dst_cluster)
         self.srcCluster = str(row.src_cluster)
-        self.srcCategory = Category(row.src_category)
         self.srcProperties = ClusterSummary(row.src_properties.no_addresses,
                                             row.src_properties.total_received,
                                             row.src_properties.total_spent)
@@ -290,6 +322,11 @@ class ClusterIncomingRelations(object):
                            round(row.value.eur, 2),
                            round(row.value.usd, 2)).__dict__
         self.noTransactions = row.no_transactions
+        self.srcBalance = compute_balance(row.src_properties.total_received,
+                                          row.src_properties.total_spent,
+                                          exchange_rate)
+        self.srcTotalReceived = compute_exchanged_value(row.src_properties.total_received,
+                                                        exchange_rate)
 
     def id(self):
         return self.srcCluster
@@ -299,8 +336,8 @@ class ClusterIncomingRelations(object):
                 "nodeType": "cluster" if self.id().isdigit() else "address",
                 "received": self.srcProperties.totalReceived,
                 "balance": (self.srcProperties.totalReceived -
-                            self.srcProperties.totalSpent),  # satoshi
-                "category": self.srcCategory.name}
+                            self.srcProperties.totalSpent)  # satoshi
+                }
         return node
 
     def toJsonEdge(self):
@@ -310,12 +347,21 @@ class ClusterIncomingRelations(object):
                 "estimatedValue": self.value}
         return edge
 
+    def toJson(self):
+        return {
+            "id": self.id(),
+            "nodeType": "cluster" if self.id().isdigit() else 'address',
+            "received": self.srcTotalReceived.__dict__,
+            "balance": self.srcBalance.__dict__,
+            "noTransactions": self.noTransactions,
+            "estimatedValue": self.value
+        }
+
 
 class ClusterOutgoingRelations(object):
-    def __init__(self, row):
+    def __init__(self, row, exchange_rate):
         self.srcCluster = str(row.src_cluster)
         self.dstCluster = str(row.dst_cluster)
-        self.dstCategory = Category(row.dst_category)
         self.dstProperties = ClusterSummary(row.dst_properties.no_addresses,
                                             row.dst_properties.total_received,
                                             row.dst_properties.total_spent)
@@ -323,6 +369,11 @@ class ClusterOutgoingRelations(object):
                            round(row.value.eur, 2),
                            round(row.value.usd, 2)).__dict__
         self.noTransactions = row.no_transactions
+        self.dstBalance = compute_balance(row.dst_properties.total_received,
+                                          row.dst_properties.total_spent,
+                                          exchange_rate)
+        self.dstTotalReceived = compute_exchanged_value(row.dst_properties.total_received,
+                                                        exchange_rate)
 
     def id(self):
         return self.dstCluster
@@ -333,7 +384,7 @@ class ClusterOutgoingRelations(object):
                 "received": self.dstProperties.totalReceived,
                 "balance": (self.dstProperties.totalReceived -
                             self.dstProperties.totalSpent),  # satoshi
-                "category": self.dstCategory.name}
+                }
         return node
 
     def toJsonEdge(self):
@@ -343,12 +394,15 @@ class ClusterOutgoingRelations(object):
                 "estimatedValue": self.value}
         return edge
 
-
-class Category(Enum):
-    Unknown = 0
-    Implicit = 1
-    Explicit = 2
-    Manual = 3
+    def toJson(self):
+        return {
+            "id": self.id(),
+            "nodeType": "cluster" if self.id().isdigit() else 'address',
+            "received": self.dstTotalReceived.__dict__,
+            "balance": self.dstBalance.__dict__,
+            "noTransactions": self.noTransactions,
+            "estimatedValue": self.value
+        }
 
 
 class AddressEgoNet(object):
@@ -359,20 +413,11 @@ class AddressEgoNet(object):
         self.implicitTags = implicit_tags
         self.incomingRelations = incoming_relations
         self.outgoingRelations = outgoing_relations
-        if self.explicitTags:
-            self.focusNodeCategory = Category.Explicit
-        else:
-            if self.implicitTags:
-                self.focusNodeCategory = Category.Implicit
-            else:
-                self.focusNodeCategory = Category.Unknown
-
         self.focusNode = [{"id": self.focusAddress.address,
                            "nodeType": "address",
                            "received": self.focusAddress.totalReceived["satoshi"],
                            "balance": (self.focusAddress.totalReceived["satoshi"] -
                                        self.focusAddress.totalSpent["satoshi"]),
-                           "category": self.focusNodeCategory.name
                            }]
 
     # receives a List[EgonetRelation]
@@ -430,18 +475,12 @@ class ClusterEgoNet(object):
         self.incomingRelations = incomingRelations
         self.outgoingRelations = outgoingRelations
 
-        if clusterTags:
-            self.focusNodeCategory = Category.Explicit
-        else:
-            self.focusNodeCategory = Category.Unknown
-
         self.focusNode = [{
             "id": self.focusCluster.cluster,
             "nodeType": "cluster",
             "received": self.focusCluster.totalReceived["satoshi"],
             "balance": (self.focusCluster.totalReceived["satoshi"] -
                         self.focusCluster.totalSpent["satoshi"]),
-            "category":self.focusNodeCategory.name
         }]
 
     def dedupNodes(self, clusterRelations):
@@ -511,3 +550,5 @@ class ClusterAddresses(object):
                                   row.total_spent.satoshi,
                                   exchange_rate)
         self.balance = balance.__dict__
+        self.inDegree = row.in_degree
+        self.outDegree = row.out_degree