+++ CHANGELOG.md
@@ -5,8 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-## [0.3.3] - 2018-12-06
+## [0.4.0] - 2019-02-01
+### Changed
+- Adjustments for new dashboard implementation
+- Removed `srcCategory`/`dstCategory` in address/cluster relations
+- Fixed exchange rates bug (EUR/USD swapped)
+- Run Docker container as non-root user
 
+### Added
+- Added primitive tests
+
+## [0.3.3] - 2018-12-06
 ### Changed
 - Reimplementation using Python/Flask